requirements.txt
streamlit>=1.38.0
pandas>=2.2.2
pytesseract>=0.3.13
Pillow>=10.4.0
matplotlib>=3.9.2
python-dotenv>=1.0.1
requests>=2.32.3
